Changed domain size variable name to domain_size everywhere. This is now the ONLY required input to pypfc. All other parameters have defaults. domain_size and ndiv are converted to numpy arrays of correct dtype internally

+++ examples/ex00_density_field.py
@@ -45,8 +45,8 @@ output_file      = 'pypfc_setup.txt'         # Output file name
 
 # Define the computational grid
 # =============================
-domain_size = params['alat'] * np.array([32, 1, 1], dtype=float)   # Domain size along the x, y and z axes
-ndiv        = 16 * np.array([32, 1, 1], dtype=int)                 # Number of grid divisions along the x, y and z axes       
+domain_size = params['alat'] * np.array([32, 1, 1]) # Domain size along the x, y and z axes
+ndiv        = 16 * np.array([32, 1, 1])             # Number of grid divisions along the x, y and z axes       
 print(f'ndiv:        {ndiv}')
 print(f'ddiv:        {domain_size / ndiv} [a]')
 print(f'domain_size: {domain_size} [a]')
